+++ src/compiler.c
@@ -1031,6 +1031,8 @@ static Py68Status py68_compile_expression(Py68Allocator *allocator,
         case PY68_TOKEN_LESS_EQUAL: return py68_emit_op(allocator, code, OP_LESS_EQUAL);
         case PY68_TOKEN_GREATER: return py68_emit_op(allocator, code, OP_GREATER);
         case PY68_TOKEN_GREATER_EQUAL: return py68_emit_op(allocator, code, OP_GREATER_EQUAL);
+        case PY68_TOKEN_IN: return py68_emit_op(allocator, code, OP_CONTAINS);
+        case PY68_TOKEN_NOT_IN: return py68_emit_op(allocator, code, OP_NOT_CONTAINS);
         default: return PY68_STATUS_SOURCE_ERROR;
         }
     }